Add user deletion, update security, and enhance migrations
Includes user deletion API, refined authentication security, and improved role-based access control across endpoints. Also enhances data seeding with user generation and aligns migrations with database updates.

+++ app/api/routers/ingredients_router.py
@@ -1,6 +1,7 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 
 from app.api.controllers import IngredientController
+from app.api.models import User
 
 from app.api.schemas.ingredients_schemas import (
     IngredientCreateSchema,
@@ -9,6 +10,7 @@ from app.api.schemas.ingredients_schemas import (
     IngredientListSchema,
     IngredientReadSchema,
 )
+from app.core.utils.security import get_current_user
 
 router = APIRouter(
     prefix="/ingredients",
@@ -23,8 +25,14 @@ router = APIRouter(
 )
 async def get_all_ingredients(
     params: IngredientListQuery = Depends(),
+    current_user: User = Depends(get_current_user),
     ingredient_controller: IngredientController = Depends(),
 ) -> IngredientListSchema:
+    if current_user.role_id not in (1, 2, 3):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this resource.",
+        )
     return await ingredient_controller.get_all_ingredients(
         payload=params,
     )
@@ -37,8 +45,14 @@ async def get_all_ingredients(
 )
 async def get_ingredient(
     ingredient_id: int,
+    current_user: User = Depends(get_current_user),
     ingredient_controller: IngredientController = Depends(),
 ) -> IngredientReadSchema:
+    if current_user.role_id not in (1, 2, 3):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this resource.",
+        )
     return await ingredient_controller.get_ingredient(
         ingredient_id=ingredient_id,
     )
@@ -49,8 +63,14 @@ async def get_ingredient(
 )
 async def create_ingredient(
     payload: IngredientCreateSchema,
+    current_user: User = Depends(get_current_user),
     ingredient_controller: IngredientController = Depends(),
 ) -> IngredientReadSchema:
+    if current_user.role_id not in (1, 2, 3):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this resource.",
+        )
     return await ingredient_controller.create_ingredient(
         payload=payload,
     )
@@ -64,8 +84,14 @@ async def create_ingredient(
 async def update_ingredient(
     ingredient_id: int,
     payload: IngredientUpdateSchema,
+    current_user: User = Depends(get_current_user),
     ingredient_controller: IngredientController = Depends(),
 ) -> IngredientReadSchema:
+    if current_user.role_id not in (1, 2, 3):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this resource.",
+        )
     return await ingredient_controller.update_ingredient(
         ingredient_id=ingredient_id,
         payload=payload,
@@ -76,8 +102,15 @@ async def update_ingredient(
     "/{ingredient_id}/", status_code=status.HTTP_204_NO_CONTENT, response_model=None
 )
 async def delete_ingredient(
-    ingredient_id: int, ingredient_controller: IngredientController = Depends()
+    ingredient_id: int,
+    current_user: User = Depends(get_current_user),
+    ingredient_controller: IngredientController = Depends(),
 ):
+    if current_user.role_id not in (1, 2, 3):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this resource.",
+        )
     return await ingredient_controller.delete_ingredient(
         ingredient_id=ingredient_id,
     )
